Initialize unit tests and Travis CI integration

+++ src/Luhn.php
@@ -40,7 +40,7 @@ class Luhn
             // Add the current digit
             $sum += $numbers[$i];
             // If the digit is even, add it again. Adjust for digits 10+ by subtracting 9.
-            ($odd_length == ($i % 2)) ? ($numbers[$i] > 4) ? ($sum += ($numbers[$i] - 9)) : ($sum += $numbers[$i]) : false;
+            ($odd_length == ($i % 2)) ? ($numbers[$i] > 4) ? ($sum += ((int)$numbers[$i] - 9)) : ($sum += $numbers[$i]) : false;
         }
         return (10 - ($sum % 10)) % 10;
     }
@@ -59,7 +59,7 @@ class Luhn
         $parity = $numDigits % 2;
 
         for ($i = $numDigits; $i >= 0; $i--) {
-            $digit = substr($number, $i, 1);
+            $digit = (int)substr($number, $i, 1);
             if (!$parity == ($i % 2)) {
                 $digit <<= 1;
             }
